splitting of "src/api/pipeline.py" into smaller sub units
