+++ src/api/pipeline_support/__init__.py
@@ -1,0 +1,1 @@
+"""Helpers for the /pipeline API route."""
